check if public key exists

+++ lib/facter/backuppc_pubkey_rsa.rb
@@ -1,5 +1,7 @@
 Facter.add('backuppc_pubkey_rsa') do
   setcode do
-    File.open('/var/lib/backuppc/.ssh/id_rsa.pub').read.split(' ')[1]
+    if File.exists?('/var/lib/backuppc/.ssh/id_rsa.pub')
+      File.open('/var/lib/backuppc/.ssh/id_rsa.pub').read.split(' ')[1]
+    end
   end
 end
